# github-crawler-api/swagger/swagger.json
{"swagger":"2.0","info":{"title":"Visualization API for github crawler","description":"API for retrieving specific data from the github crawler","contact":{"name":"Rick Proost, Wim Spaargaren, Vincent Robbemond","email":"rpjproost@gmail.com","url":"http://127.0.0.1"},"version":""},"host":"86.87.235.82:8081","basePath":"/v1","schemes":["http"],"consumes":["application/json","application/x-www-form-urlencoded"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/repositories/list":{"get":{"tags":["repositories"],"summary":"list repositories","description":"Retrieve all repositories in DB","operationId":"repositories#list","produces":["application/vnd.goa.error","application/vnd.repository+json; type=collection"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RepositoryCollection"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/repositories/{repoID}":{"get":{"tags":["repositories"],"summary":"show repositories","description":"Retrieve repository from db","operationId":"repositories#show","produces":["application/vnd.repository+json","application/vnd.goa.error"],"parameters":[{"name":"repoID","in":"path","description":"Repository ID","required":true,"type":"integer","minimum":1}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Repository"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/repositories/{repoID}/commits/list":{"get":{"tags":["commits"],"summary":"list commits","description":"Retrieve commits between timespan with users","operationId":"commits#list","produces":["application/vnd.goa.error","application/vnd.commit+json; type=collection"],"parameters":[{"name":"from","in":"query","description":"From date","required":false,"type":"string"},{"name":"limit","in":"query","description":"Limit the results","required":false,"type":"integer"},{"name":"repoID","in":"path","description":"Repository ID","required":true,"type":"integer","minimum":1},{"name":"until","in":"query","description":"Till ID","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/CommitCollection"}},"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/repositories/{repoID}/commits/{sha}":{"get":{"tags":["commits"],"summary":"show commits","description":"Retrieve commit from db","operationId":"commits#show","produces":["application/vnd.goa.error","application/vnd.commit+json"],"parameters":[{"name":"repoID","in":"path","description":"Repository ID","required":true,"type":"integer","minimum":1},{"name":"sha","in":"path","description":"SHA of the commit","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Commit"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}}},"definitions":{"Commit":{"title":"Mediatype identifier: application/vnd.commit+json; view=default","type":"object","properties":{"authorOfCommit":{"$ref":"#/definitions/Ghuser"},"id":{"type":"integer","description":"ID of the commit in the database","example":9037984394037558769,"format":"int64"},"message":{"type":"string","description":"Message of the commit","example":"Sunt asperiores dolor necessitatibus."},"sha":{"type":"string","description":"Unique identifier of the commit","example":"Suscipit consectetur sed soluta natus doloribus velit."},"timestamp":{"type":"string","description":"Time the commit happened","example":"2006-05-20T18:06:34+02:00","format":"date-time"}},"description":"Commit data (default view)","example":{"authorOfCommit":{"id":5430185467799328504,"locationForUser":{"id":1140970880266293086,"lat":0.566770896282312,"lng":0.45074023793163837},"login":"Sit et esse maiores sit.","type":"Consequatur eos atque similique."},"id":9037984394037558769,"message":"Sunt asperiores dolor necessitatibus.","sha":"Suscipit consectetur sed soluta natus doloribus velit.","timestamp":"2006-05-20T18:06:34+02:00"},"required":["id","sha","authorOfCommit","timestamp"]},"CommitCollection":{"title":"Mediatype identifier: application/vnd.commit+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Commit"},"description":"CommitCollection is the media type for an array of Commit (default view)","example":[{"authorOfCommit":{"id":5430185467799328504,"locationForUser":{"id":1140970880266293086,"lat":0.566770896282312,"lng":0.45074023793163837},"login":"Sit et esse maiores sit.","type":"Consequatur eos atque similique."},"id":9037984394037558769,"message":"Sunt asperiores dolor necessitatibus.","sha":"Suscipit consectetur sed soluta natus doloribus velit.","timestamp":"2006-05-20T18:06:34+02:00"}]},"Ghuser":{"title":"Mediatype identifier: application/vnd.ghuser+json; view=default","type":"object","properties":{"id":{"type":"integer","description":"ID of the user in the database","example":5430185467799328504,"format":"int64"},"locationForUser":{"$ref":"#/definitions/Location"},"login":{"type":"string","description":"Unique username of the user","example":"Sit et esse maiores sit."},"type":{"type":"string","description":"Type of the user","example":"Consequatur eos atque similique."}},"description":"GH user data (default view)","example":{"id":5430185467799328504,"locationForUser":{"id":1140970880266293086,"lat":0.566770896282312,"lng":0.45074023793163837},"login":"Sit et esse maiores sit.","type":"Consequatur eos atque similique."},"required":["id","login","type"]},"Location":{"title":"Mediatype identifier: application/vnd.location+json; view=default","type":"object","properties":{"id":{"type":"integer","description":"ID of the location in the database","example":1140970880266293086,"format":"int64"},"lat":{"type":"number","description":"coordinates lat","example":0.566770896282312,"format":"double"},"lng":{"type":"number","description":"coordinates lng","example":0.45074023793163837,"format":"double"}},"description":"Location as geocoordinates (default view)","example":{"id":1140970880266293086,"lat":0.566770896282312,"lng":0.45074023793163837},"required":["id","lat","lng"]},"Repository":{"title":"Mediatype identifier: application/vnd.repository+json; view=default","type":"object","properties":{"first_commit":{"$ref":"#/definitions/Commit"},"full_name":{"type":"string","description":"Full name of the repo","example":"Dolor ut."},"id":{"type":"integer","description":"ID of the commit in the database","example":6858816164497510772,"format":"int64"},"org":{"type":"boolean","description":"If owner is an organization","example":true},"owner":{"type":"string","description":"Name of the owner of the repository","example":"Non magnam."},"project_id":{"type":"number","description":"Time the commit happened","example":0.11750343986984432,"format":"double"}},"description":"Commit data (default view)","example":{"first_commit":{"authorOfCommit":{"id":5430185467799328504,"locationForUser":{"id":1140970880266293086,"lat":0.566770896282312,"lng":0.45074023793163837},"login":"Sit et esse maiores sit.","type":"Consequatur eos atque similique."},"id":9037984394037558769,"message":"Sunt asperiores dolor necessitatibus.","sha":"Suscipit consectetur sed soluta natus doloribus velit.","timestamp":"2006-05-20T18:06:34+02:00"},"full_name":"Dolor ut.","id":6858816164497510772,"org":true,"owner":"Non magnam.","project_id":0.11750343986984432},"required":["id","owner","full_name","project_id"]},"RepositoryCollection":{"title":"Mediatype identifier: application/vnd.repository+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Repository"},"description":"RepositoryCollection is the media type for an array of Repository (default view)","example":[{"first_commit":{"authorOfCommit":{"id":5430185467799328504,"locationForUser":{"id":1140970880266293086,"lat":0.566770896282312,"lng":0.45074023793163837},"login":"Sit et esse maiores sit.","type":"Consequatur eos atque similique."},"id":9037984394037558769,"message":"Sunt asperiores dolor necessitatibus.","sha":"Suscipit consectetur sed soluta natus doloribus velit.","timestamp":"2006-05-20T18:06:34+02:00"},"full_name":"Dolor ut.","id":6858816164497510772,"org":true,"owner":"Non magnam.","project_id":0.11750343986984432},{"first_commit":{"authorOfCommit":{"id":5430185467799328504,"locationForUser":{"id":1140970880266293086,"lat":0.566770896282312,"lng":0.45074023793163837},"login":"Sit et esse maiores sit.","type":"Consequatur eos atque similique."},"id":9037984394037558769,"message":"Sunt asperiores dolor necessitatibus.","sha":"Suscipit consectetur sed soluta natus doloribus velit.","timestamp":"2006-05-20T18:06:34+02:00"},"full_name":"Dolor ut.","id":6858816164497510772,"org":true,"owner":"Non magnam.","project_id":0.11750343986984432},{"first_commit":{"authorOfCommit":{"id":5430185467799328504,"locationForUser":{"id":1140970880266293086,"lat":0.566770896282312,"lng":0.45074023793163837},"login":"Sit et esse maiores sit.","type":"Consequatur eos atque similique."},"id":9037984394037558769,"message":"Sunt asperiores dolor necessitatibus.","sha":"Suscipit consectetur sed soluta natus doloribus velit.","timestamp":"2006-05-20T18:06:34+02:00"},"full_name":"Dolor ut.","id":6858816164497510772,"org":true,"owner":"Non magnam.","project_id":0.11750343986984432}]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"NoContent":{"description":"No Content"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/Commit"}}}}